+++ tailwind.config.ts
@@ -321,6 +321,7 @@ export const Catppuccin = {
 };
 
 const config: Config = {
+  darkMode: "class",
   content: [
     "./pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./components/**/*.{js,ts,jsx,tsx,mdx}",